+++ public/js/blocks/daily-top.js
@@ -1,6 +1,6 @@
 import { frameHeader } from "../components/frame-header.js";
 import { webtoonLongCard } from "../components/webtoon-long-card.js";
-import { webtoons } from "../data/webtoons.js";
+import { webtoons } from "../../../data/webtoons.js";
 
 const dailyWebtoons = webtoons.dailyTop;
