+++ public/js/setting-contents.js
@@ -1,4 +1,4 @@
-import { webtoons } from "./data/webtoons.js";
+import { webtoons } from "../../data/webtoons.js";
 import { renderingContents } from "./rendering/contents.js";
 import { promotionBanner } from "./components/promotion-banner.js";
 import { subAdBanner } from "./components/sub-ad-banner.js";
@@ -26,6 +26,7 @@ const mainNavContents = {
       frame: "unuse",
     },
   ],
+  웹툰: [],
 };
 
 const nav = [
